moved fit utils from utils to respective submodules
