+++ src/napari_stress/_spherical_harmonics/fit_utils.py
@@ -1,8 +1,5 @@
-import numpy as np
-import inspect
-from napari.types import PointsData
-
 from .._stress import lebedev_info_SPB as lebedev_info
+import numpy as np
 
 def Least_Squares_Harmonic_Fit(
     fit_degree: int, sample_locations: tuple, values: np.ndarray
